respect the BUILD_SHARED_LIBS flag
this is a standard CMake flag that
is supposed to control whether libraries
are compiled as shared or static by
default.
this commit uses that flag to set
the default values for whether
each MMG library is shared or static,
while users can still override those
defaults if they want.