Modif of delauney kernel
Before commit, we were skipping an edge if she belongs to a tetra with a xtetra but not to a boundary face (because it is not possible to say if the edge is boundary or not from only this tet). Now, edge shell is computed and the edge is skipped only if she is boundary, which is consistent with pattern kernel impelmentation.
It also makes sense to have an edge treatment that doesn't depend to the data structure.