Fix incomplete (directional) test in MMG3D_chkmeshedgestags
If, when first meet, the edge had a lower tag in the tetra than the tag found from one of the next tetra, the tag inconsistency was not detected due to the tag value accumulation inside the hash table.

If the first stored tag of an edge was higher than one of the tag found after, the inconsistency was suitabily detected.